Cast inline-size sampler total to bigint to fix Rust decode mismatch
`pg_total_relation_size()` returns BIGINT, but `SUM(BIGINT)` in
PostgreSQL is promoted to NUMERIC to avoid overflow. The final
addition of two NUMERIC subqueries also stays NUMERIC. The Rust side
binds to `i64` (BIGINT/INT8), which fails to decode NUMERIC and trips
`mismatched types; Rust type 'i64' is not compatible with SQL type 'NUMERIC'`
on every sampler poll.

Cast the outer addition to BIGINT in SQL so the column type matches
the i64 destination. Total inline byte counts cap at i64::MAX
(~9.2 EB), so this can't overflow in any realistic deployment.

+++ etl-destinations/src/ducklake/inline_size.rs
@@ -109,8 +109,8 @@ fn pending_inline_data_bytes_query(metadata_schema: &str) -> String {
              ) AS total_bytes
          )
          SELECT
-             (SELECT total_bytes FROM inline_data_bytes)
-             + (SELECT total_bytes FROM inline_delete_bytes);"
+             ((SELECT total_bytes FROM inline_data_bytes)
+              + (SELECT total_bytes FROM inline_delete_bytes))::bigint;"
     )
 }
